Fix final test (hopefully :) )
